add synth borrow
